+++ js/variables.js
@@ -1,0 +1,7 @@
+const AVATARS = [];
+const OFFER_TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const OFFER_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const numberOfOffers = 10;
+const numberOfAvatars = 10;
+
+export {AVATARS, OFFER_TYPE, OFFER_FEATURES, numberOfAvatars, numberOfOffers};
